[Bug]: Failure because of missing index log should be more helpful Fixes #7
